Move insertation of eGon100RE capacities to ScenarioCapacities
