+++ src/egon/data/datasets/pypsaeursec/__init__.py
@@ -116,160 +116,6 @@ def run_pypsa_eur_sec():
     )
 
 
-def eGon100_capacities():
-    """Inserts installed capacities for the eGon100 scenario
-
-    Returns
-    -------
-    None.
-
-    """
-    # read-in installed capacities
-    execute_pypsa_eur_sec = False
-    cwd = Path(".")
-
-    if execute_pypsa_eur_sec:
-        filepath = cwd / "run-pypsa-eur-sec"
-        pypsa_eur_sec_repos = filepath / "pypsa-eur-sec"
-        # Read YAML file
-        pes_egonconfig = pypsa_eur_sec_repos / "config_egon.yaml"
-        with open(pes_egonconfig, "r") as stream:
-            data_config = yaml.safe_load(stream)
-
-        target_file = (
-            pypsa_eur_sec_repos
-            / "results"
-            / data_config["run"]
-            / "csvs"
-            / "nodal_capacities.csv"
-        )
-
-    else:
-        target_file = (
-            cwd
-            / "data_bundle_egon_data"
-            / "pypsa_eur_sec"
-            / "2021-egondata-integration"
-            / "csvs"
-            / "nodal_capacities.csv"
-        )
-
-    df = pd.read_csv(target_file, skiprows=5)
-    df.columns = ["component", "country", "carrier", "p_nom"]
-
-    df.set_index("carrier", inplace=True)
-
-    df = df[df.country.str[:2] == "DE"]
-
-    # Drop country column
-    df.drop("country", axis=1, inplace=True)
-
-    # Drop copmponents which will be optimized in eGo
-    unused_carrier = [
-        "BEV charger",
-        "DAC",
-        "H2 Electrolysis",
-        "electricity distribution grid",
-        "home battery charger",
-        "home battery discharger",
-        "H2",
-        "Li ion",
-        "home battery",
-        "residential rural water tanks charger",
-        "residential rural water tanks discharger",
-        "services rural water tanks charger",
-        "services rural water tanks discharger",
-        "residential rural water tanks",
-        "services rural water tanks",
-        "urban central water tanks",
-        "urban central water tanks charger",
-        "urban central water tanks discharger",
-        "H2 Fuel Cell",
-    ]
-
-    df = df[~df.index.isin(unused_carrier)]
-
-    df.index = df.index.str.replace(" ", "_")
-
-    # Aggregate offshore wind
-    df = df.append(
-        pd.DataFrame(
-            index=["wind_offshore"],
-            data={
-                "p_nom": (df.p_nom["offwind-ac"] + df.p_nom["offwind-dc"]),
-                "component": df.component["offwind-ac"],
-            },
-        )
-    )
-    df = df.drop(["offwind-ac", "offwind-dc"])
-
-    # Aggregate technologies with and without carbon_capture (CC)
-    for carrier in ["SMR", "urban_central_gas_CHP"]:
-        df.p_nom[carrier] += df.p_nom[f"{carrier}_CC"]
-        df = df.drop([f"{carrier}_CC"])
-
-    # Aggregate residential and services rural heat supply
-    for merge_carrier in [
-        "rural_resistive_heater",
-        "rural_ground_heat_pump",
-        "rural_gas_boiler",
-        "rural_solar_thermal",
-    ]:
-        df = df.append(
-            pd.DataFrame(
-                index=[merge_carrier],
-                data={
-                    "p_nom": (
-                        df.p_nom[f"residential_{merge_carrier}"]
-                        + df.p_nom[f"services_{merge_carrier}"]
-                    ),
-                    "component": df.component[f"residential_{merge_carrier}"],
-                },
-            )
-        )
-        df = df.drop(
-            [f"residential_{merge_carrier}", f"services_{merge_carrier}"]
-        )
-
-    # Rename carriers
-    df.rename(
-        {
-            "onwind": "wind_onshore",
-            "ror": "run_of_river",
-            "PHS": "pumped_hydro",
-            "OCGT": "gas",
-            "rural_ground_heat_pump": "rural_heat_pump",
-        },
-        inplace=True,
-    )
-
-    # Reset index
-    df = df.reset_index()
-
-    # Rename columns
-    df.rename(
-        {"p_nom": "capacity", "index": "carrier"}, axis="columns", inplace=True
-    )
-
-    df["scenario_name"] = "eGon100RE"
-    df["nuts"] = "DE"
-
-    db.execute_sql(
-        """
-        DELETE FROM supply.egon_scenario_capacities
-        WHERE scenario_name='eGon100RE'
-        """
-    )
-
-    df.to_sql(
-        "egon_scenario_capacities",
-        schema="supply",
-        con=db.engine(),
-        if_exists="append",
-        index=False,
-    )
-
-
 def neighbor_reduction():
 
     # Set execute_pypsa_eur_sec to False until optional task is implemented
@@ -954,16 +800,16 @@ def neighbor_reduction():
 execute_pypsa_eur_sec = False
 
 if execute_pypsa_eur_sec:
-    tasks = (run_pypsa_eur_sec, {eGon100_capacities, neighbor_reduction})
+    tasks = (run_pypsa_eur_sec, neighbor_reduction)
 else:
-    tasks = {eGon100_capacities, neighbor_reduction}
+    tasks = neighbor_reduction
 
 
 class PypsaEurSec(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="PypsaEurSec",
-            version="0.0.1",
+            version="0.0.2",
             dependencies=dependencies,
             tasks=tasks,
         )
